Feat: FooterList TypeScript 적용
- 기존 코드 리터럴 템플릿으로 변경
- 타입스크립트 적용
